requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
beautifulsoup4==4.13.4
    # via bs4
bs4==0.0.2
    # via -r requirements.in
certifi==2025.4.26
    # via requests
charset-normalizer==3.4.2
    # via requests
django==5.2.1
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.16.0
    # via -r requirements.in
idna==3.10
    # via requests
requests==2.32.3
    # via -r requirements.in
soupsieve==2.7
    # via beautifulsoup4
sqlparse==0.5.3
    # via django
typing-extensions==4.14.0
    # via beautifulsoup4
tzdata==2025.2
    # via django
urllib3==2.4.0
    # via requests
